+++ src/txstack.h
@@ -52,7 +52,7 @@ class Stack {
 
 private:
     size_t                 last_xid; // id of the last
-    int                    xid_boot;
+    size_t                 xid_boot;
     Istring                cur_lid;    // the id to be pushed on uids[]
     std::vector<AttList>   attributes; // the main table of attributes
     std::vector<Xml *>     enames;     // the main table of element names
